+++ core/src/com/unciv/logic/city/CityConstructions.kt
@@ -71,6 +71,13 @@ class CityConstructions {
         val stats = Stats()
         for (building in getBuiltBuildings())
             stats.add(building.getStats(cityInfo.civInfo))
+
+        for (unique in builtBuildingUniqueMap.getAllUniques()) when (unique.placeholderText) {
+            "[] Per [] Population in this city" -> stats.add(Stats.parse(unique.params[0]).times(cityInfo.population.population / unique.params[1].toFloat()))
+            "[] once [] is discovered" -> if (cityInfo.civInfo.tech.isResearched(unique.params[1])) stats.add(Stats.parse(unique.params[0]))
+        }
+
+        // This is to be deprecated and converted to "[stats] Per [N] Population in this city" - keeping it here to that mods with this can still work for now
         stats.science += (builtBuildingUniqueMap.getAllUniques().count { it.text == "+1 Science Per 2 Population" } * cityInfo.population.population / 2).toFloat()
         return stats
     }
@@ -123,7 +130,7 @@ class CityConstructions {
         if (currentConstructionSnapshot!=""
                 && !PerpetualConstruction.perpetualConstructionsMap.containsKey(currentConstructionSnapshot)) {
             val turnsLeft = turnsToConstruction(currentConstructionSnapshot)
-            result += " - $turnsLeft ${Fonts.turn}"
+            result += " - $turnsLeft${Fonts.turn}"
         }
         return result
     }
@@ -389,9 +396,15 @@ class CityConstructions {
         if (isQueueFull()) return
         if (currentConstructionFromQueue == "" || currentConstructionFromQueue == "Nothing") {
             currentConstructionFromQueue = constructionName
-            currentConstructionIsUserSet = true
+        } else if (getConstruction(constructionQueue.last()) is PerpetualConstruction) {
+            if (getConstruction(constructionName) is PerpetualConstruction) {  // perpetual constructions will replace each other
+                constructionQueue.removeAt(constructionQueue.size - 1)
+                constructionQueue.add(constructionName)
+            } else
+                constructionQueue.add(constructionQueue.size - 1, constructionName) // insert new construction before perpetual one
         } else
             constructionQueue.add(constructionName)
+        currentConstructionIsUserSet = true
     }
 
     /** If this was done automatically, we should automatically try to choose a new construction and treat it as such */