+++ core/src/com/unciv/ui/newgamescreen/MapOptionsTable.kt
@@ -145,7 +145,11 @@ class MapOptionsTable(val newGameScreen: NewGameScreen): Table() {
         }
         mapFileSelectBox.items = mapFiles
         val selectedItem = mapFiles.firstOrNull { it.fileHandle.name()==mapParameters.name }
-        if(selectedItem!=null) mapFileSelectBox.selected = selectedItem
+        if (selectedItem != null) mapFileSelectBox.selected = selectedItem
+        else if (!mapFiles.isEmpty) {
+            mapFileSelectBox.selected = mapFiles.first()
+            newGameScreen.gameSetupInfo.mapFile = mapFileSelectBox.selected.fileHandle
+        }
 
         mapFileSelectBox.onChange {
             val mapFile =  mapFileSelectBox.selected.fileHandle