+++ tests/src/com/unciv/testing/BasicTests.kt
@@ -76,16 +76,8 @@ class BasicTests {
 
         UncivGame.Current = UncivGame("")
         UncivGame.Current.settings = GameSettings().apply { language = "Italian" }
-        val x = "+1 Gold, +2 Production".tr()
-        print(x)
     }
 
-//    @Test // This should NOT run as part of the test suite!
-//    fun tryUnzip(){
-//        Zip.extractFolder("""C:\Users\LENOVO\Downloads\Rebuild.rar""")
-//        Zip.downloadAndExtract("/Mods/Reasoures.zip", FileHandle("""C:\Users\LENOVO\Downloads"""))
-//        Zip.downloadAndExtract("https://github.com/yairm210/Unciv-IV-mod/archive/master.zip", FileHandle("""C:\Users\LENOVO\Downloads"""))
-//    }
 
 //    @Test // This should NOT run as part of the test suite!
 //    fun tryGetGithubTopicInfo(){